Added another Wav
This is a frequency sweep, so we should see it sweep accross the
visualizer, but this isn't happening, I do, however, recognize the issue
(i hope).  Daniel, I tried to do this myself before I left home, and
I'll do it if I get a moment before you see this, but you should do a
couple of things

--Do not display values[0]
--try to use values[0] to normalize our max.  The DC(that's waht this is
called) seems pretty huge compared to all the other values

Thanks!